backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Web Automation & Browser Control
playwright==1.41.0
selenium==4.17.2
webdriver-manager==4.0.1

# HTTP & Requests
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Scheduling & Background Tasks
APScheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Image Processing & Screenshots
Pillow==10.2.0
opencv-python==4.9.0.80

# AI & ML
google-generativeai==0.3.2
openai==1.10.0

# Notifications & Alerts
python-telegram-bot==20.7
slack-sdk==3.26.2
sendgrid==6.11.0
twilio==8.11.1

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.40.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
pytz==2023.3
pandas==2.2.0
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.6.0

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CRM Integrations
hubspot-api-client==8.2.1
simple-salesforce==1.12.5

# Performance & Analytics
locust==2.20.0
beautifulsoup4==4.12.3
lxml==5.1.0

# CORS & Middleware
python-cors==1.0.0
slowapi==0.1.9

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
ipython==8.20.0
